feat: Raiden shoot-em-up, audio system, QScript v3 builtins, docs alignment
**New game: examples/raiden/** — vertical STG / shoot-em-up
  - 8-directional movement (held keys), 6-level weapon (Vulcan→5-way→triple laser)
  - Bomb system (screen-clear + boss damage + flash overlay)
  - Enemy formations (sine-curve dive), 4 enemy types, ground turrets
  - End-of-stage boss with aimed 3-way fan fire
  - Power-up drops: P (weapon), B (bomb), medal (score)
  - Parallax starfield, scrolling terrain (land/water alternation)
  - Auto-fire, enemy fire (turret aimed at player, bomber straight drop)
  - All sprites procedural 32×32 (triangles + radial gradients, CC0)
  - 6 chiptune WAVs (CC0: music loop, shoot/hit/explode/death/powerup)

**New package: internal/audio/**
  - Minimal WAV decoder (PCM 8/16-bit), Player interface, StdoutSink
  - afplay/aplay/paplay backend with temp-file WAV for macOS support
  - nopSink default (headless tests quiet); cmd/qorm init() registers StdoutSink

**QScript v3 builtins**
  - sin/cos/tan/atan2/sqrt (curved motion, aimed shots in games)
  - playSound/playMusic/stopMusic (audio routing through expr→runtime adapter)
  - Previously added (v2): typeof/isArray/isString/isNumber/isBool/isObject/isNull,
    jsonEncode/jsonDecode, flatten, push/unshift/pop/shift, reverse/sort,
    indexOf/includes, charAt/substring/repeat/padStart/padEnd/trimStart/trimEnd

**Canvas engine fixes**
  - Board root with explicit size for game scenes (fixes fill-height clipping of
    absolutely-positioned children — the 'white block' bug)
  - imageSrc / MeasureImage / RecordImage thread list-scope vars through
    evalPropStrWithVars so {{item}}-bound src works inside gridview/list renderItem
  - SoftwareRenderer clipBounds nil-safety (empty clips→img.Bounds())

**Mario v2** — 60fps physics, AABB per-axis collision, board+camera,
  hold-to-move (keys+keyReleases), variable-height jump, goomba AI,
  power-ups (mushroom/fire), particles, bump blocks, multi-screen level

**Docs alignment**
  - Fixed broken append()→push() example in expressions.md (EN+ZH)
  - Added 26 missing builtins to expressions.md docs (trig, type checks, JSON, audio, call)
  - Capability count: ~26→43 in native-middlelayer.md (EN+ZH)
  - Website games page metadata updated (Tetris & 2048 → Tetris, 2048 & Raiden)
  - Main page game links updated (EN+ZH); comparison table rendering card +canvas
  - gen-templates.mjs catalog + templates.js regenerated with raiden (27 templates)

**Test infrastructure**
  - mario_script_test.go removed (replaced by mario_v2_test.go)
  - New: raiden_test.go (6 tests), raiden_pos_test.go (regression),
    raiden_perf_test.go (benchmark: 1.3ms/frame, 764fps headless),
    audio_adapter_test.go, audio_test.go, playback_test.go

+++ internal/render/canvas/style.go
@@ -108,7 +108,14 @@ type NodeStyle struct {
 	// instead of flowing it — the coordinate model of an infinite-canvas
 	// board. Out of flow, so it neither consumes flex space nor reflows
 	// siblings.
-	PosX, PosY   int
+	//
+	// Stored as float64 (sub-pixel): a 60fps physics tick (mario) wants
+	// `x: state.mario * 32` to land between two pixels so the sprite moves
+	// smoothly, not in 1-pixel snaps. The renderer rounds to int at the
+	// final draw (measure.go); float here is the contract, int is the
+	// pixel. Old examples that pass integer x/y keep integer values
+	// end-to-end — the change is a strict superset.
+	PosX, PosY   float64
 	HasPos       bool
 	Align        string
 	AlignSelf    string // CSS align-self (style/layout alignSelf) — distinct from Align (align-items)
@@ -160,8 +167,8 @@ func (s *NodeStyle) scaleBy(f int) {
 	s.BoxShadowY *= f
 	s.Width *= f
 	s.Height *= f
-	s.PosX *= f
-	s.PosY *= f
+	s.PosX *= float64(f)
+	s.PosY *= float64(f)
 	s.FontSize *= f
 	s.BorderRadius *= float64(f)
 	s.StrokeWidth *= float64(f)
@@ -662,24 +669,26 @@ func applyStyleProps(s *NodeStyle, style map[string]any, rt *runtime.Runtime, sc
 	}
 
 	// Absolute position: x/y (native) or left/top (HTML alias). Either key
-	// present marks the node positioned; the missing axis reads 0.
+	// present marks the node positioned; the missing axis reads 0. Stored
+	// as float64 (see PosX/PosY doc) so a 60fps physics tick can land
+	// between pixels; the renderer rounds at draw time.
 	for _, key := range []string{"x", "left"} {
 		v := esp(style[key])
 		if f, ok := v.(float64); ok {
-			s.PosX = int(f)
+			s.PosX = f
 			s.HasPos = true
 		} else if i, ok := v.(int); ok {
-			s.PosX = i
+			s.PosX = float64(i)
 			s.HasPos = true
 		}
 	}
 	for _, key := range []string{"y", "top"} {
 		v := esp(style[key])
 		if f, ok := v.(float64); ok {
-			s.PosY = int(f)
+			s.PosY = f
 			s.HasPos = true
 		} else if i, ok := v.(int); ok {
-			s.PosY = i
+			s.PosY = float64(i)
 			s.HasPos = true
 		}
 	}
